+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The following changes have been implemented but not released yet:
 
 ## [Unreleased]
 
+### New features
+
+- Manage public keys attached to your profile with the functions `addJwkToJwks`,
+  `addPublicKeyToProfileJwks`, `getProfileJwksIri` and `setProfileJwks`, from the
+  `@inrupt/solid-client/profile/jwks` module.
+
 The following sections document changes that have been released already:
 
 ## [1.11.1] - 2021-09-02